Module pallet_dapp_staking_v3::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

claim_bonus_reward

Used to claim bonus reward for a smart contract, if eligible.

claim_dapp_reward

Used to claim dApp reward for the specified era.

claim_staker_rewards

Claims some staker rewards, if user has any. In the case of a successful call, at least one era will be claimed, with the possibility of multiple claims happening.

claim_unlocked

Claims all of fully unlocked chunks, removing the lock from them.

cleanup_expired_entries

Cleanup expired stake entries for the contract.

fix_account

A call used to fix accounts with inconsistent state, where frozen balance is actually higher than what’s available.

force

Used to force a change of era or subperiod. The effect isn’t immediate but will happen on the next block.

lock

Locks additional funds into dApp staking.

maintenance_mode

Used to enable or disable maintenance mode. Can only be called by manager origin.

register

Used to register a new contract for dApp staking.

relock_unlocking

Warning: Doc-Only

set_dapp_owner

Used to change dApp owner.

set_dapp_reward_beneficiary

Used to modify the reward beneficiary account for a dApp.

stake

Stake the specified amount on a smart contract. The precise amount specified must be available for staking. The total amount staked on a dApp must be greater than the minimum required value.

unbond_and_unstake

Wrapper around legacy-like unbond_and_unstake.

unlock

Attempts to start the unlocking process for the specified amount.

unregister

Unregister dApp from dApp staking protocol, making it ineligible for future rewards. This doesn’t remove the dApp completely from the system just yet, but it can no longer be used for staking.

unstake

Unstake the specified amount from a smart contract. The amount specified must not exceed what’s staked, otherwise the call will fail.

unstake_from_unregistered

Used to unstake funds from a contract that was unregistered after an account staked on it. This is required if staker wants to re-stake these funds on another active contract during the ongoing period.

withdraw_unbonded

Wrapper around legacy-like withdraw_unbonded.