Trait pallet_dapp_staking_v3::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent> + TryInto<Event<Self>>;
    type RuntimeFreezeReason: From<FreezeReason>;
    type Currency: FunMutateFreeze<Self::AccountId, Id = Self::RuntimeFreezeReason, Balance = Balance>;
    type SmartContract: Parameter + Member + MaxEncodedLen + SmartContractHandle<Self::AccountId>;
    type ManagerOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin>;
    type NativePriceProvider: PriceProvider;
    type StakingRewardHandler: StakingRewardHandler<Self::AccountId>;
    type CycleConfiguration: CycleConfiguration;
    type Observers: DAppStakingObserver;
    type AccountCheck: AccountCheck<Self::AccountId>;
    type EraRewardSpanLength: Get<u32>;
    type RewardRetentionInPeriods: Get<PeriodNumber>;
    type MaxNumberOfContracts: Get<u32>;
    type MaxUnlockingChunks: Get<u32>;
    type MinimumLockedAmount: Get<Balance>;
    type UnlockingPeriod: Get<EraNumber>;
    type MaxNumberOfStakedContracts: Get<u32>;
    type MinimumStakeAmount: Get<Balance>;
    type NumberOfTiers: Get<u32>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent> + TryInto<Event<Self>>

The overarching event type.

type RuntimeFreezeReason: From<FreezeReason>

The overarching freeze reason.

type Currency: FunMutateFreeze<Self::AccountId, Id = Self::RuntimeFreezeReason, Balance = Balance>

Currency used for staking. Reference: https://github.com/paritytech/substrate/pull/12951/

type SmartContract: Parameter + Member + MaxEncodedLen + SmartContractHandle<Self::AccountId>

Describes smart contract in the context required by dApp staking.

type ManagerOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin>

Privileged origin for managing dApp staking pallet.

type NativePriceProvider: PriceProvider

Used to provide price information about the native token.

type StakingRewardHandler: StakingRewardHandler<Self::AccountId>

Used to handle reward payouts & reward pool amount fetching.

type CycleConfiguration: CycleConfiguration

Describes era length, subperiods & period length, as well as cycle length.

type Observers: DAppStakingObserver

dApp staking event observers, notified when certain events occur.

type AccountCheck: AccountCheck<Self::AccountId>

Used to check whether an account is allowed to participate in dApp staking.

type EraRewardSpanLength: Get<u32>

Maximum length of a single era reward span length entry.

type RewardRetentionInPeriods: Get<PeriodNumber>

Number of periods for which we keep rewards available for claiming. After that period, they are no longer claimable.

type MaxNumberOfContracts: Get<u32>

Maximum number of contracts that can be integrated into dApp staking at once.

type MaxUnlockingChunks: Get<u32>

Maximum number of unlocking chunks that can exist per account at a time.

type MinimumLockedAmount: Get<Balance>

Minimum amount an account has to lock in dApp staking in order to participate.

type UnlockingPeriod: Get<EraNumber>

Number of standard eras that need to pass before unlocking chunk can be claimed. Even though it’s expressed in ‘eras’, it’s actually measured in number of blocks.

type MaxNumberOfStakedContracts: Get<u32>

Maximum amount of stake contract entries an account is allowed to have at once.

type MinimumStakeAmount: Get<Balance>

Minimum amount staker can stake on a contract.

type NumberOfTiers: Get<u32>

Number of different tiers.

type WeightInfo: WeightInfo

Weight info for various calls & operations in the pallet.