Type Alias pallet_dapp_staking_v3::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn unbond_and_unstake( origin: OriginFor<T>, _contract_id: T::SmartContract, value: Balance ) -> DispatchResult

Wrapper around legacy-like unbond_and_unstake.

Used to support legacy Ledger users so they can start the unlocking process for their funds.

pub fn withdraw_unbonded(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Wrapper around legacy-like withdraw_unbonded.

Used to support legacy Ledger users so they can reclaim unlocked chunks back into their transferable free balance.

pub fn maintenance_mode(origin: OriginFor<T>, enabled: bool) -> DispatchResult

Used to enable or disable maintenance mode. Can only be called by manager origin.

pub fn register( origin: OriginFor<T>, owner: T::AccountId, smart_contract: T::SmartContract ) -> DispatchResult

Used to register a new contract for dApp staking.

If successful, smart contract will be assigned a simple, unique numerical identifier. Owner is set to be initial beneficiary & manager of the dApp.

pub fn set_dapp_reward_beneficiary( origin: OriginFor<T>, smart_contract: T::SmartContract, beneficiary: Option<T::AccountId> ) -> DispatchResult

Used to modify the reward beneficiary account for a dApp.

Caller has to be dApp owner. If set to None, rewards will be deposited to the dApp owner. After this call, all existing & future rewards will be paid out to the beneficiary.

pub fn set_dapp_owner( origin: OriginFor<T>, smart_contract: T::SmartContract, new_owner: T::AccountId ) -> DispatchResult

Used to change dApp owner.

Can be called by dApp owner or dApp staking manager origin. This is useful in two cases:

when the dApp owner account is compromised, manager can change the owner to a new account

if project wants to transfer ownership to a new account (DAO, multisig, etc.).

pub fn unregister( origin: OriginFor<T>, smart_contract: T::SmartContract ) -> DispatchResult

Unregister dApp from dApp staking protocol, making it ineligible for future rewards. This doesn’t remove the dApp completely from the system just yet, but it can no longer be used for staking.

Can be called by dApp staking manager origin.

pub fn lock(origin: OriginFor<T>, amount: Balance) -> DispatchResultWithPostInfo

Locks additional funds into dApp staking.

In case caller account doesn’t have sufficient balance to cover the specified amount, everything is locked. After adjustment, lock amount must be greater than zero and in total must be equal or greater than the minimum locked amount.

Locked amount can immediately be used for staking.

pub fn unlock(origin: OriginFor<T>, amount: Balance) -> DispatchResult

Attempts to start the unlocking process for the specified amount.

Only the amount that isn’t actively used for staking can be unlocked. If the amount is greater than the available amount for unlocking, everything is unlocked. If the remaining locked amount would take the account below the minimum locked amount, everything is unlocked.

pub fn claim_unlocked(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Claims all of fully unlocked chunks, removing the lock from them.

pub fn relock_unlocking(origin: OriginFor<T>) -> DispatchResult

pub fn stake( origin: OriginFor<T>, smart_contract: T::SmartContract, amount: Balance ) -> DispatchResult

Stake the specified amount on a smart contract. The precise amount specified must be available for staking. The total amount staked on a dApp must be greater than the minimum required value.

Depending on the period type, appropriate stake amount will be updated. During Voting subperiod, voting stake amount is updated, and same for Build&Earn subperiod.

Staked amount is only eligible for rewards from the next era onwards.

pub fn unstake( origin: OriginFor<T>, smart_contract: T::SmartContract, amount: Balance ) -> DispatchResult

Unstake the specified amount from a smart contract. The amount specified must not exceed what’s staked, otherwise the call will fail.

If unstaking the specified amount would take staker below the minimum stake threshold, everything is unstaked.

Depending on the period type, appropriate stake amount will be updated. In case amount is unstaked during Voting subperiod, the voting amount is reduced. In case amount is unstaked during Build&Earn subperiod, first the build_and_earn is reduced, and any spillover is subtracted from the voting amount.

pub fn claim_staker_rewards(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Claims some staker rewards, if user has any. In the case of a successful call, at least one era will be claimed, with the possibility of multiple claims happening.

pub fn claim_bonus_reward( origin: OriginFor<T>, smart_contract: T::SmartContract ) -> DispatchResult

Used to claim bonus reward for a smart contract, if eligible.

pub fn claim_dapp_reward( origin: OriginFor<T>, smart_contract: T::SmartContract, era: EraNumber ) -> DispatchResult

Used to claim dApp reward for the specified era.

pub fn unstake_from_unregistered( origin: OriginFor<T>, smart_contract: T::SmartContract ) -> DispatchResult

Used to unstake funds from a contract that was unregistered after an account staked on it. This is required if staker wants to re-stake these funds on another active contract during the ongoing period.

pub fn cleanup_expired_entries( origin: OriginFor<T> ) -> DispatchResultWithPostInfo

Cleanup expired stake entries for the contract.

Entry is considered to be expired if:

It’s from a past period & the account wasn’t a loyal staker, meaning there’s no claimable bonus reward.

It’s from a period older than the oldest claimable period, regardless whether the account was loyal or not.

pub fn force(origin: OriginFor<T>, forcing_type: ForcingType) -> DispatchResult

Used to force a change of era or subperiod. The effect isn’t immediate but will happen on the next block.

Used for testing purposes, when we want to force an era change, or a subperiod change. Not intended to be used in production, except in case of unforeseen circumstances.

Can only be called by the root origin.

pub fn claim_staker_rewards_for( origin: OriginFor<T>, account: T::AccountId ) -> DispatchResultWithPostInfo

Claims some staker rewards for the specified account, if they have any. In the case of a successful call, at least one era will be claimed, with the possibility of multiple claims happening.

pub fn claim_bonus_reward_for( origin: OriginFor<T>, account: T::AccountId, smart_contract: T::SmartContract ) -> DispatchResult

Used to claim bonus reward for a smart contract on behalf of the specified account, if eligible.

pub fn fix_account( origin: OriginFor<T>, account: T::AccountId ) -> DispatchResultWithPostInfo

A call used to fix accounts with inconsistent state, where frozen balance is actually higher than what’s available.

The approach is as simple as possible:

Caller provides an account to fix.

If account is eligible for the fix, all unlocking chunks are modified to be claimable immediately.

The claim_unlocked call is executed using the provided account as the origin.

All states are updated accordingly, and the account is no longer in an inconsistent state.

The benchmarked weight of the claim_unlocked call is used as a base, and additional overestimated weight is added. Call doesn’t touch any storage items that aren’t already touched by the claim_unlocked call, hence the simplified approach.

impl<T: Config> Pallet<T>

pub fn is_staker(account: &T::AccountId) -> bool

true if the account is a staker, false otherwise.

pub fn era_reward_span_index(era: EraNumber) -> EraNumber

Calculates the EraRewardSpan index for the specified era.

pub fn unlocking_period() -> BlockNumber

Unlocking period expressed in the number of blocks.

pub fn get_dapp_tier_assignment() -> BTreeMap<DAppId, RankedTier>

Returns the dApp tier assignment for the current era, based on the current stake amounts.