Type Alias pallet_dapp_staking_v3::AccountLedgerFor

pub type AccountLedgerFor<T> = AccountLedger<<T as Config>::MaxUnlockingChunks>;

Aliased Type

struct AccountLedgerFor<T> {
    pub locked: u128,
    pub unlocking: BoundedVec<UnlockingChunk, <T as Config>::MaxUnlockingChunks>,
    pub staked: StakeAmount,
    pub staked_future: Option<StakeAmount>,
    pub contract_stake_count: u32,
}

Fields
locked: u128

How much active locked amount an account has. This can be used for staking.

unlocking: BoundedVec<UnlockingChunk, <T as Config>::MaxUnlockingChunks>

Vector of all the unlocking chunks. This is also considered locked but cannot be used for staking.

staked: StakeAmount

Primary field used to store how much was staked in a particular era.

staked_future: Option<StakeAmount>

Secondary field used to store ‘stake’ information for the ‘next era’. This is needed since stake amount is only applicable from the next era after it’s been staked.

Both stake and staked_future must ALWAYS refer to the same period. If staked_future is Some, it will always be EXACTLY one era after the staked field era.

contract_stake_count: u32

Number of contract stake entries in storage.

Implementations

impl<UnlockingLen> AccountLedger<UnlockingLen>where UnlockingLen: Get<u32>,

pub fn is_empty(&self) -> bool

Empty if no locked/unlocking/staked info exists.

pub fn active_locked_amount(&self) -> Balance

Returns active locked amount. If zero, means that associated account hasn’t got any active locked funds.

It is possible that some funds are undergoing the unlocking period, but they aren’t considered active in that case.

pub fn unlocking_amount(&self) -> Balance

Returns unlocking amount. If zero, means that associated account hasn’t got any unlocking chunks.

pub fn total_locked_amount(&self) -> Balance

Total locked amount by the user. Includes both active locked amount & unlocking amount.

pub fn add_lock_amount(&mut self, amount: Balance)

Adds the specified amount to the total locked amount.

pub fn subtract_lock_amount(&mut self, amount: Balance)

Subtracts the specified amount of the total locked amount.

pub fn add_unlocking_chunk( &mut self, amount: Balance, unlock_block: BlockNumber ) -> Result<(), AccountLedgerError>

Adds the specified amount to the unlocking chunks.

If entry for the specified block already exists, it’s updated.

If entry for the specified block doesn’t exist, it’s created and insertion is attempted. In case vector has no more capacity, error is returned, and whole operation is a noop.

pub fn unlockable_amount(&self, current_period: PeriodNumber) -> Balance

Amount available for unlocking.

pub fn claim_unlocked(&mut self, current_block_number: BlockNumber) -> Balance

Claims all of the fully unlocked chunks, and returns the total claimable amount.

pub fn consume_unlocking_chunks(&mut self) -> Balance

Consumes all of the unlocking chunks, and returns the total amount being unlocked.

pub fn stakeable_amount(&self, active_period: PeriodNumber) -> Balance

Amount that is available for staking.

This is equal to the total active locked amount, minus the staked amount already active.

pub fn staked_amount(&self, active_period: PeriodNumber) -> Balance

Amount that is staked, in respect to the currently active period.

pub fn staked_amount_for_type( &self, subperiod: Subperiod, period: PeriodNumber ) -> Balance

How much is staked for the specified subperiod, in respect to the specified era.

pub fn add_stake_amount( &mut self, amount: Balance, current_era: EraNumber, current_period_info: PeriodInfo ) -> Result<(), AccountLedgerError>

Adds the specified amount to total staked amount, if possible.

Staking can only be done for the ongoing period, and era.

The period requirement enforces staking in the ongoing period.

The era requirement enforces staking in the ongoing era.

The 2nd condition is needed to prevent stakers from building a significant history of stakes, without claiming the rewards. So if a historic era exists as an entry, stakers will first need to claim the pending rewards, before they can stake again.

Additionally, the staked amount must not exceed what’s available for staking.

pub fn unstake_amount( &mut self, amount: Balance, current_era: EraNumber, current_period_info: PeriodInfo ) -> Result<(), AccountLedgerError>

Subtracts the specified amount from the total staked amount, if possible.

Unstake can only be called if the entry for the current era exists. In case historic entry exists, rewards first need to be claimed, before unstaking is possible. Similar as with stake functionality, this is to prevent staker from building a significant history of stakes.

pub fn staked_period(&self) -> Option<PeriodNumber>

Period for which account has staking information or None if no staking information exists.

pub fn earliest_staked_era(&self) -> Option<EraNumber>

Earliest era for which the account has staking information or None if no staking information exists.

pub fn maybe_cleanup_expired( &mut self, valid_threshold_period: PeriodNumber ) -> bool

Cleanup staking information if it has expired.

Args

valid_threshold_period - last period for which entries can still be considered valid.

true if any change was made, false otherwise.

pub fn claim_up_to_era( &mut self, era: EraNumber, period_end: Option<EraNumber> ) -> Result<EraStakePairIter, AccountLedgerError>

‘Claim’ rewards up to the specified era. Returns an iterator over the (era, amount) pairs, where amount describes the staked amount eligible for reward in the appropriate era.

If period_end is provided, it’s used to determine whether all applicable chunks have been claimed.