Type Alias pallet_dapp_staking_v3::DAppTierRewardsFor

pub type DAppTierRewardsFor<T> = DAppTierRewards<<T as Config>::MaxNumberOfContracts, <T as Config>::NumberOfTiers>;

Aliased Type

struct DAppTierRewardsFor<T> {
    pub dapps: BoundedBTreeMap<u16, RankedTier, <T as Config>::MaxNumberOfContracts>,
    pub rewards: BoundedVec<u128, <T as Config>::NumberOfTiers>,
    pub period: u32,
    pub rank_rewards: BoundedVec<u128, <T as Config>::NumberOfTiers>,
}

Fields
dapps: BoundedBTreeMap<u16, RankedTier, <T as Config>::MaxNumberOfContracts>

DApps and their corresponding tiers (or None if they have been claimed in the meantime)

rewards: BoundedVec<u128, <T as Config>::NumberOfTiers>

Rewards for each tier. First entry refers to the first tier, and so on.

period: u32

Period during which this struct was created.

rank_rewards: BoundedVec<u128, <T as Config>::NumberOfTiers>

Rank reward for each tier. First entry refers to the first tier, and so on.